third_party/blink/web_tests/platform/generic/fast/dom/CSSStyleDeclaration/css-computed-style-item-expected.txt

This tests indexing outside the range of the computed style object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof computedStyle.length is "number"
PASS computedStyle[computedStyle.length] is undefined.
PASS computedStyle[-1] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
